cloud-function (2)/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery-storage
pandas
db-dtypes
